+++ flutter-hms-adsprime/android/src/main/java/com/huawei/hms/flutter/ads/HmsAdsPlugin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -73,7 +73,7 @@ import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.platform.PlatformViewRegistry;
 
 public class HmsAdsPlugin implements FlutterPlugin, ActivityAware, MethodCallHandler {
-    private static final String TAG = "HUAWEI_ADS_PLUGIN";
+    private static final String TAG = "HUAWEI_ADS_PRIME_PLUGIN";
     private Activity activity;
     private Context context;
     private MethodChannel methodChannel;