+++ flutter-hms-adsprime/android/src/main/java/com/huawei/hms/flutter/ads/adslite/nativead/NativeAdController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ import androidx.annotation.NonNull;
 
 import com.huawei.hms.ads.AdListener;
 import com.huawei.hms.ads.AdParam;
+import com.huawei.hms.ads.AppInfo;
 import com.huawei.hms.ads.nativead.DislikeAdReason;
 import com.huawei.hms.ads.nativead.NativeAd;
 import com.huawei.hms.ads.nativead.NativeAdConfiguration;
@@ -32,6 +33,7 @@ import com.huawei.hms.flutter.ads.logger.HMSLogger;
 import com.huawei.hms.flutter.ads.utils.FromMap;
 import com.huawei.hms.flutter.ads.utils.ToMap;
 import com.huawei.hms.flutter.ads.utils.constants.ErrorCodes;
+import com.huawei.openalliance.ad.beans.metadata.PromoteInfo;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -43,15 +45,23 @@ import io.flutter.plugin.common.MethodChannel;
 
 public class NativeAdController implements MethodChannel.MethodCallHandler {
     private static final String TAG = "NativeAdController";
+
     private String id;
+
     private MethodChannel channel;
+
     private Context context;
 
     private NativeAd nativeAd;
+
     private AdListener adListener;
+
     private NativeAd.NativeAdLoadedListener nativeAdLoadedListener;
+
     private NativeAdLoader nativeAdLoader;
+
     private String adSlotId;
+
     private Map<String, Object> adParam;
 
     NativeAdController(String id, MethodChannel channel, Context context) {
@@ -125,6 +135,75 @@ public class NativeAdController implements MethodChannel.MethodCallHandler {
                 result.success(true);
                 HMSLogger.getInstance(context).sendSingleEvent("gotoWhyThisAdPage");
                 break;
+            case "showAppDetailPage":
+                HMSLogger.getInstance(context).startMethodExecutionTimer("showAppDetailPage");
+                try {
+                    nativeAd.showAppDetailPage(context);
+                    HMSLogger.getInstance(context).sendSingleEvent("showAppDetailPage");
+                    result.success(null);
+                } catch (Exception e) {
+                    Log.e(TAG, e.getMessage());
+                    HMSLogger.getInstance(context).sendSingleEvent("showAppDetailPage", ErrorCodes.INNER);
+                    result.error(ErrorCodes.INNER, "showAppDetailPage failed.", e.getMessage());
+                }
+                break;
+            case "getPromoteInfo":
+                HMSLogger.getInstance(context).startMethodExecutionTimer("getPromoteInfo");
+                PromoteInfo promoteInfo = nativeAd.getPromoteInfo();
+                if (promoteInfo == null) {
+                    HMSLogger.getInstance(context).sendSingleEvent("getPromoteInfo", "-1");
+                    result.error("-1", "PromoteInfo is null.", "");
+                } else {
+                    Map<String, Object> map = new HashMap<>();
+                    map.put("promoteType", promoteInfo.getType());
+                    map.put("promoteName", promoteInfo.getName());
+                    HMSLogger.getInstance(context).sendSingleEvent("getPromoteInfo");
+                    result.success(map);
+                }
+                break;
+            case "getAppInfo":
+                HMSLogger.getInstance(context).startMethodExecutionTimer("getAppInfo");
+                AppInfo appInfo = nativeAd.getAppInfo();
+                if (appInfo == null) {
+                    HMSLogger.getInstance(context).sendSingleEvent("getAppInfo", "-1");
+                    result.error("-1", "AppInfo is null.", "");
+                } else {
+                    Map<String, Object> map = new HashMap<>();
+                    map.put("appName", appInfo.getAppName());
+                    map.put("appDesc", appInfo.getAppDesc());
+                    map.put("iconUrl", appInfo.getAppIconUrl());
+                    map.put("packageName", appInfo.getPkgName());
+                    map.put("privacyLink", appInfo.getPrivacyLink());
+                    map.put("appDetailUrl", appInfo.getAppDetailUrl());
+                    map.put("versionName", appInfo.getVersionName());
+                    map.put("developerName", appInfo.getDeveloperName());
+                    HMSLogger.getInstance(context).sendSingleEvent("getAppInfo");
+                    result.success(map);
+                }
+                break;
+            case "showPrivacyPolicy":
+                HMSLogger.getInstance(context).startMethodExecutionTimer("showPrivacyPolicy");
+                try {
+                    nativeAd.getAppInfo().showPrivacyPolicy(context);
+                    result.success(null);
+                } catch (Exception e) {
+                    Log.e(TAG, e.getMessage());
+                    HMSLogger.getInstance(context).sendSingleEvent("showPrivacyPolicy", ErrorCodes.INNER);
+                    result.error(ErrorCodes.INNER, "showPrivacyPolicy failed.", e.getMessage());
+                }
+                break;
+            case "showPermissionPage":
+                HMSLogger.getInstance(context).startMethodExecutionTimer("showPermissionPage");
+                try {
+                    nativeAd.getAppInfo().showPermissionPage(context);
+                    HMSLogger.getInstance(context).sendSingleEvent("showPermissionPage");
+                    result.success(null);
+                } catch (Exception e) {
+                    Log.e(TAG, e.getMessage());
+                    HMSLogger.getInstance(context).sendSingleEvent("showPermissionPage", ErrorCodes.INNER);
+                    result.error(ErrorCodes.INNER, "showPermissionPage failed.", e.getMessage());
+                }
+                break;
             default:
                 onNativeGetterMethodCall(call, result);
         }
@@ -177,23 +256,21 @@ public class NativeAdController implements MethodChannel.MethodCallHandler {
                 break;
             case "hasVideo":
                 HMSLogger.getInstance(context).startMethodExecutionTimer("hasVideo");
-                boolean hasVideo = nativeAd.getVideoOperator() != null
-                && nativeAd.getVideoOperator().hasVideo();
+                boolean hasVideo = nativeAd.getVideoOperator() != null && nativeAd.getVideoOperator().hasVideo();
                 result.success(hasVideo);
                 HMSLogger.getInstance(context).sendSingleEvent("hasVideo");
                 break;
             case "isCustomOperateEnabled":
                 HMSLogger.getInstance(context).startMethodExecutionTimer("isCustomOperateEnabled");
-                boolean isCustomOperateEnabled = nativeAd.getVideoOperator() != null
-                        && nativeAd.getVideoOperator().isCustomizeOperateEnabled();
-                        result.success(isCustomOperateEnabled);
+                boolean isCustomOperateEnabled = nativeAd.getVideoOperator() != null && nativeAd.getVideoOperator()
+                    .isCustomizeOperateEnabled();
+                result.success(isCustomOperateEnabled);
                 HMSLogger.getInstance(context).sendSingleEvent("isCustomOperateEnabled");
                 break;
             case "isMuted":
                 HMSLogger.getInstance(context).startMethodExecutionTimer("isMuted");
-                boolean isMuted = nativeAd.getVideoOperator() != null
-                        && nativeAd.getVideoOperator().isMuted();
-                        result.success(isMuted);
+                boolean isMuted = nativeAd.getVideoOperator() != null && nativeAd.getVideoOperator().isMuted();
+                result.success(isMuted);
                 HMSLogger.getInstance(context).sendSingleEvent("isMuted");
                 break;
             case "mute":
@@ -236,7 +313,8 @@ public class NativeAdController implements MethodChannel.MethodCallHandler {
             result.success(true);
             HMSLogger.getInstance(context).sendSingleEvent("muteNativeAd");
         } else {
-            result.error(ErrorCodes.NULL_PARAM, "Video Operator or boolean parameter is null. Mute failed. | isMute : " + mute, "");
+            result.error(ErrorCodes.NULL_PARAM,
+                "Video Operator or boolean parameter is null. Mute failed. | isMute : " + mute, "");
             HMSLogger.getInstance(context).sendSingleEvent("muteNativeAd");
         }
     }
@@ -295,9 +373,7 @@ public class NativeAdController implements MethodChannel.MethodCallHandler {
                 if (adListener != null) {
                     builder.setAdListener(adListener);
                 }
-                nativeAdLoader = builder
-                    .setNativeAdOptions(generateNativeAdConfiguration(adConfigurationMap))
-                    .build();
+                nativeAdLoader = builder.setNativeAdOptions(generateNativeAdConfiguration(adConfigurationMap)).build();
                 if (nativeAd == null || slotIdChanged) {
                     adParam = adParamMap;
                     loadAd();
@@ -306,7 +382,8 @@ public class NativeAdController implements MethodChannel.MethodCallHandler {
             result.success(true);
             HMSLogger.getInstance(context).sendSingleEvent("setupNativeAdController");
         } else {
-            result.error(ErrorCodes.NULL_PARAM, "adSlotId is either null or empty. Controller setup failed. | Controller id : " + id, "");
+            result.error(ErrorCodes.NULL_PARAM,
+                "adSlotId is either null or empty. Controller setup failed. | Controller id : " + id, "");
             HMSLogger.getInstance(context).sendSingleEvent("setupNativeAdController");
         }
     }