+++ flutter-hms-adsprime/example/lib/pages/ads_menu_page.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -30,7 +30,7 @@ class AdsMenuPage extends StatelessWidget {
       appBar: AppBar(
         backgroundColor: Colors.redAccent,
         title: const Text(
-          'Huawei Ads Plugin Demo',
+          'Huawei Ads Prime Plugin Demo',
           style: TextStyle(
             color: Colors.white,
           ),