+++ flutter-hms-adsprime/example/lib/pages/publisher/native_ad_page.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -77,6 +77,12 @@ class _NativeAdPageState extends State<NativeAdPage> {
 
     bool? isTransparencyOpen = await controller.isTransparencyOpen();
     debugPrint('isTransparencyOpen : $isTransparencyOpen');
+
+    PromoteInfo? promoteInfo = await controller.getPromoteInfo();
+    debugPrint('getPromoteInfo : ${promoteInfo.toString()}');
+
+    AppInfo? appInfo = await controller.getAppInfo();
+    debugPrint('getAppInfo: ${appInfo.toString()}');
   }
 
   @override