+++ flutter-hms-adsprime/lib/huawei_adsprime.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -42,6 +42,8 @@ part 'src/adslite/nativead/native_ad.dart';
 part 'src/adslite/nativead/native_ad_configuration.dart';
 part 'src/adslite/nativead/native_ad_controller.dart';
 part 'src/adslite/nativead/native_styles.dart';
+part 'src/adslite/nativead/app_info.dart';
+part 'src/adslite/nativead/promote_info.dart';
 part 'src/adslite/request_options.dart';
 part 'src/adslite/reward/reward_ad.dart';
 part 'src/adslite/reward/reward_verify_config.dart';