+++ flutter-hms-adsprime/lib/src/adslite/nativead/native_ad_controller.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -204,6 +204,26 @@ class NativeAdController {
       responseList.add(DislikeAdReason(reason));
     }
     return responseList;
+  }
+
+  Future<void> showAppDetailPage() async {
+    await _channel.invokeMethod('showAppDetailPage');
+  }
+
+  Future<PromoteInfo> getPromoteInfo() async {
+    return PromoteInfo.fromJson(await _channel.invokeMethod('getPromoteInfo'));
+  }
+
+  Future<AppInfo> getAppInfo() async {
+    return AppInfo.fromJson(await _channel.invokeMethod('getAppInfo'));
+  }
+
+  Future<void> showPrivacyPolicy() async {
+    await _channel.invokeMethod('showPrivacyPolicy');
+  }
+
+  Future<void> showPermissionPage() async {
+    await _channel.invokeMethod('showPermissionPage');
   }
 
   Future<bool?> destroy() {